hello-world/requirements.txt
python-dotenv==1.0.0
pytest==8.0.0
pytest-cov==4.1.0
flake8==7.0.0
bandit==1.7.7